live-transcribe-backend/live-stt-service/requirement.txt
fastapi
uvicorn
deepgram-sdk==3.9.0
websockets==10.4
python-dotenv
